structure changed, tailwind configured
