+++ src/App.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import './App.css';
-import Navbar from './components/Navbar';
-import {BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Home from './components/pages/Home';
+import React from "react";
+import "./App.css";
+import Navbar from "./components/Navbar/Navbar";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import Home from "./components/Home/pages/Home";
 
 function App() {
   return (
     <>
-    <Router>
-      <Navbar />
-      <Switch>
-        <Route path='/' exact component={Home} />
-      </Switch>
-    </Router>
+      <Router>
+        <Navbar />
+        <Switch>
+          <Route path="/" exact component={Home} />
+        </Switch>
+      </Router>
     </>
   );
 }
