+++ src/components/Home/Assets/CSS/Slideshow.css
