+++ src/components/Home/pages/Home.js
@@ -1,0 +1,21 @@
+import React from "react";
+import "../../../App.css";
+// import Cards from "../Cards";
+import Opening from "../components/Opening";
+import Slideshow from "../components/Slideshow";
+import Footer from "../components/Footer";
+import SignUp from "../components/SignUp";
+
+function Home() {
+  return (
+    <>
+      <Opening />
+      {/* <Cards /> */}
+      <Slideshow />
+      <SignUp />
+      <Footer />
+    </>
+  );
+}
+
+export default Home;
